+++ jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-	testRegex: 'src/test/test-.*\\.js',
-};