+++ src/test/test-UA-no-WebRTC.ts
@@ -1,27 +1,28 @@
 /* eslint no-console: 0*/
 
-require('./include/common');
-const testUA = require('./include/testUA');
-const JsSIP = require('../..');
+import './include/common';
+import * as consts from './include/consts';
+
+// eslint-disable-next-line @typescript-eslint/no-require-imports
+const JsSIP = require('../JsSIP.js');
+const { UA, WebSocketInterface, Exceptions, C } = JsSIP;
 
 describe('UA No WebRTC', () => {
 	test('UA wrong configuration', () => {
-		expect(() => new JsSIP.UA({ lalala: 'lololo' })).toThrow(
-			JsSIP.Exceptions.ConfigurationError
+		expect(() => new UA({ lalala: 'lololo' } as never)).toThrow(
+			Exceptions.ConfigurationError
 		);
 	});
 
 	test('UA no WS connection', () => {
-		const config = testUA.UA_CONFIGURATION;
-		const wsSocket = new JsSIP.WebSocketInterface(
-			testUA.SOCKET_DESCRIPTION.url
-		);
+		const config = consts.UA_CONFIGURATION;
+		const wsSocket = new WebSocketInterface(consts.SOCKET_DESCRIPTION['url']);
 
-		config.sockets = wsSocket;
+		config['sockets'] = wsSocket;
 
-		const ua = new JsSIP.UA(config);
+		const ua = new UA(config);
 
-		expect(ua instanceof JsSIP.UA).toBeTruthy();
+		expect(ua instanceof UA).toBeTruthy();
 
 		ua.start();
 
@@ -43,19 +44,19 @@ describe('UA No WebRTC', () => {
 			'<sip:anonymous@anonymous.invalid;transport=ws;ob>'
 		);
 
-		for (const parameter in testUA.UA_CONFIGURATION_AFTER_START) {
+		for (const parameter in consts.UA_CONFIGURATION_AFTER_START) {
 			if (
 				Object.prototype.hasOwnProperty.call(
-					testUA.UA_CONFIGURATION_AFTER_START,
+					consts.UA_CONFIGURATION_AFTER_START,
 					parameter
 				)
 			) {
 				switch (parameter) {
 					case 'uri':
 					case 'registrar_server': {
+						// eslint-disable-next-line jest/no-conditional-expect
 						expect(ua.configuration[parameter].toString()).toBe(
-							testUA.UA_CONFIGURATION_AFTER_START[parameter],
-							`testing parameter ${parameter}`
+							consts.UA_CONFIGURATION_AFTER_START[parameter]
 						);
 						break;
 					}
@@ -64,17 +65,17 @@ describe('UA No WebRTC', () => {
 						break;
 					}
 					default: {
+						// eslint-disable-next-line jest/no-conditional-expect
 						expect(ua.configuration[parameter]).toBe(
-							testUA.UA_CONFIGURATION_AFTER_START[parameter],
-							`testing parameter ${parameter}`
+							consts.UA_CONFIGURATION_AFTER_START[parameter]
 						);
 					}
 				}
 			}
 		}
 
-		const transport = testUA.UA_TRANSPORT_AFTER_START;
-		const sockets = transport.sockets;
+		const transport = consts.UA_TRANSPORT_AFTER_START;
+		const sockets = transport['sockets'];
 		const socket = sockets[0].socket;
 
 		expect(sockets.length).toEqual(ua.transport.sockets.length);
@@ -83,12 +84,15 @@ describe('UA No WebRTC', () => {
 		expect(socket.sip_uri).toEqual(ua.transport.sip_uri);
 		expect(socket.url).toEqual(ua.transport.url);
 
-		expect(transport.recovery_options).toEqual(ua.transport.recovery_options);
+		expect(transport['recovery_options']).toEqual(
+			ua.transport.recovery_options
+		);
 
 		ua.sendMessage('test', 'FAIL WITH CONNECTION_ERROR PLEASE', {
 			eventHandlers: {
-				failed: function (e) {
-					expect(e.cause).toEqual(JsSIP.C.causes.CONNECTION_ERROR);
+				// eslint-disable-next-line @typescript-eslint/no-explicit-any
+				failed: function (e: any) {
+					expect(e.cause).toEqual(C.causes.CONNECTION_ERROR);
 				},
 			},
 		});